testdrive: move some avro utilities into the avro utility mod
